FEAT: rename IUtil; moving fn
